+++ spb/__init__.py
@@ -7,6 +7,7 @@ from spb.functions import (
     plot3d_parametric_line,
     plot3d_parametric_surface,
     plot3d_implicit,
+    plot3d_spherical,
     plot_implicit,
     plot_polar,
     plot_geometry,